finishdm888

+++ week-04/day-02/Garden/tree.h
@@ -9,7 +9,7 @@
 class Tree {
 public:
     Tree();
-    int _waterTree;
+    float _waterTree;
     bool _treeNeedsWater = (_waterTree < 10);
 
     float _howMuchWaterIsMissingFromTree();
